1.3.0
- Added in capacity to retain the best individual from the first
generation, and then plot that with everything else at trial end.
- Added temporary function to `Ecoilogy/` to plot things when they
don’t go as planned.
- Fixed problem in genetix.py (327) where sometimes a non-existent
element of best_fitness_list would be accessed. Still not sure why this
happened, but it seems to be fixed now.
- Tweaked epsilonCalc
- Began to edit `__erf_homogeneity():` to accommodate for part of the
fitness having to do with going above a certain PPM at a range.
- Changed the way Mutation Method is handled. Instead of 4 different
Boolean values, the status is simply contained within the variable
MutationScheme. Just change the value of this to a string associated
with a mutation method to alter performance.
- Began building scheduling module.
- `evolver` now terminates based on an arbitrary value of epsilon. In
the future, should add this value to myconst.